requirements.txt
torch
torchvision
transformers
flask
Pillow
